refactor: cleanup for deployment
