+++ src/components/ClassMembers.tsx
@@ -1,6 +1,6 @@
 import { Link } from 'react-router-dom';
-import { DocumentationClass } from '../typings/Docs';
-import { scopedName } from '../util/scopedName';
+import type { DocumentationClass } from '../typings/Docs';
+import { scopedName } from '../util';
 import { CodeBlock } from './CodeBlock';
 import { ParameterTable } from './ParameterTable';
