--- src/components/GettingStarted.tsx
@@ -1,7 +1,0 @@
-import { useParams } from 'react-router-dom';
-
-export function GettingStarted() {
-	const { topic } = useParams<{ topic: string }>();
-
-	return <>Topic: {topic}</>;
-}
