+++ src/components/SidebarItems.tsx
@@ -1,11 +1,11 @@
 import { Link } from 'react-router-dom';
-import { useTypedSelector } from '../store/Hooks';
-import { useGetDocsQuery } from '../store/DocsSlice';
+import { useGetDocsQuery } from '../store/DocsApi';
 import { type SourceIdType, Sources } from '../data';
+import { useTypedSelector } from '../store/Store';
 
 export function SidebarItems() {
 	const { selectedSourceId, selectedVersion } = useTypedSelector(state => state.docsSettings);
-	const { data, error, isLoading } = useGetDocsQuery({
+	const { data, error, isLoading, isSuccess } = useGetDocsQuery({
 		docsRepo: Sources[selectedSourceId as SourceIdType].docsRepo,
 		sourceId: selectedSourceId,
 		version: selectedVersion,
@@ -16,7 +16,7 @@ export function SidebarItems() {
 		if (isLoading) return <>Loading...</>;
 		return data?.[itemType].map(item => {
 			return (
-				<li className='pt-2 pl-3 text-sm' key={item.name}>
+				<li className='pt-2 pl-3 text-base' key={item.name}>
 					<Link to={`${itemType}/${item.name}`}>{item.name}</Link>
 				</li>
 			);
@@ -25,22 +25,22 @@ export function SidebarItems() {
 
 	return (
 		<div className='text-left'>
-			<div className='getting-started'>
-				<div className='pt-6 text-lg font-bold'>Getting Started</div>
-				<ul className='space-y-1'>
-					<li className='pt-2 pl-3 text-sm'>
-						<Link to='getting-started/general'>General</Link>
-					</li>
-				</ul>
-			</div>
-			<div className='classes'>
-				<div className='pt-6 text-lg font-bold'>Classes</div>
-				<ul className='space-y-1'>{createItemList('classes')}</ul>
-			</div>
-			<div className='typedefs'>
-				<div className='pt-6 text-lg font-bold'>Typedefs</div>
-				<ul className='space-y-1'>{createItemList('typedefs')}</ul>
-			</div>
+			{isSuccess && data?.classes.length ? (
+				<div className='classes'>
+					<div className='pt-6 text-lg font-bold'>Classes</div>
+					<ul className='space-y-1'>{createItemList('classes')}</ul>
+				</div>
+			) : (
+				<></>
+			)}
+			{isSuccess && data?.typedefs.length ? (
+				<div className='typedefs'>
+					<div className='pt-6 text-lg font-bold'>Typedefs</div>
+					<ul className='space-y-1'>{createItemList('typedefs')}</ul>
+				</div>
+			) : (
+				<></>
+			)}
 		</div>
 	);
 }
