--- src/store/Hooks.ts
@@ -1,5 +1,0 @@
-import { TypedUseSelectorHook, useDispatch, useSelector } from 'react-redux';
-import type { StoreDispatchType, StoreStateType } from './Store';
-
-export const useTypedDispatch = () => useDispatch<StoreDispatchType>();
-export const useTypedSelector: TypedUseSelectorHook<StoreStateType> = useSelector;
