+++ src/store/Store.ts
@@ -1,7 +1,8 @@
 import { configureStore } from '@reduxjs/toolkit';
+import { useDispatch, type TypedUseSelectorHook, useSelector } from 'react-redux';
 import { docsSettingsReducer } from './DocsSettingsSlice';
-import { docsApi } from './DocsSlice';
-import { githubApi } from './VersionsSlice';
+import { docsApi } from './DocsApi';
+import { githubApi } from './GithubApi';
 
 export const store = configureStore({
 	reducer: {
@@ -14,3 +15,6 @@ export const store = configureStore({
 
 export type StoreStateType = ReturnType<typeof store.getState>;
 export type StoreDispatchType = typeof store.dispatch;
+
+export const useTypedDispatch = () => useDispatch<StoreDispatchType>();
+export const useTypedSelector: TypedUseSelectorHook<StoreStateType> = useSelector;
